[Fix] 문의 User , Admin 기능 분리
